@@ -1,3 +1,3 @@
 module github.com/athxx/goutils
 
-go 1.17
+go 1.19
